requirements.txt
django-bootstrap4==24.3
django-widget-tweaks==1.5.0
psycopg2==2.9.9